Change some integer types to bigger ones and refactor code
  - Change the argument revision and UID variable types to
    unsigned long long int in the PLUGIN and JOB types just
    to be on the safe side.
  - Refactor code and clean up output printed to STDOUT.
  - Update TODO.txt

+++ src/headers/plugin.h
@@ -29,7 +29,7 @@
 		unsigned int argc;
 		char *cache_path;
 		char *cache_name;
-		unsigned int arg_rev;
-		unsigned int uid;
+		unsigned long long int arg_rev;
+		unsigned long long int uid;
 	} PLUGIN;
 #endif
